python/pdf/requirements.txt
fpdf2==2.4.1
Pillow==8.2.0
pkg-resources==0.0.0
